+++ opensrs/opensrsapi.py
@@ -651,3 +651,21 @@ class OpenSRS(object):
             object='domain',
             attributes={'domain': domain_name}
         ).get_data()
+
+    def enable_domain_auto_renewal(self, cookie, domain_name):
+        self._set_domain_auto_renewal_status(cookie, domain_name, True)
+
+    def disable_domain_auto_renewal(self, cookie, domain_name):
+        self._set_domain_auto_renewal_status(cookie, domain_name, False)
+
+    def _set_domain_auto_renewal_status(self, cookie, domain_name, enabled):
+        attributes = {
+            'data': 'expire_action',
+            'auto_renew': int(enabled),
+            'let_expire': int(not enabled)
+        }
+
+        return self._req(
+            action='MODIFY', object='DOMAIN', attributes=attributes,
+            cookie=cookie
+        )